feat(frontend): share QuestPanel + 2-column holder detail (#435)
- Extract QuestPanel (with QuestCard / QuestStateBadge / questMeta) into
  `app/components/quests/QuestPanel.tsx` so duty and holder detail share
  one implementation.
- Mirror the duty detail's desktop 60/40 split on holder detail with
  the QuestPanel in the right column on desktop and stacked on mobile.
- Gate the "クエストを作成" CTA behind `canCreate` — only show it when
  the viewer actually holds role share for the hat (any wearer's shard
  on duty detail, the current wearer's shard on holder detail).
- Role card on `/{treeId}/role` now shows both `詳細` and `担当を追加`
  buttons; "詳細を見る" copy trimmed to "詳細".
